+++ R/zzz.R
@@ -24,4 +24,6 @@
 .state$gd_base_url <- "https://www.googleapis.com"
 .state$gd_base_url_files_v2 <-
   file.path(.state$gd_base_url, "drive", "v2", "files")
+.state$gd_base_url_files_v3 <-
+  file.path(.state$gd_base_url, "drive", "v3", "files")
 #.state$gs_base_url <- "https://spreadsheets.google.com/feeds"