+++ man-roxygen/return-googlesheet.R
@@ -1,0 +1,1 @@
+#' @return a \code{\link{googlesheet}} object